UI fixes

+++ client/src/features/FreezerLog.jsx
@@ -108,8 +108,8 @@ function FreezerLog() {
   };
 
   const handleDelete = async (id) => {
-    // const confirmed = window.confirm("Are you sure you want to delete this item?");
-    // if (!confirmed) return;
+    const confirmed = window.confirm("Are you sure you want to delete this item?");
+    if (!confirmed) return;
 
     try {
       await axios.delete(`${import.meta.env.VITE_API_BASE_URL}/api/freezer-items/${id}`);
